move outdated tests to new file
